+++ full-moon/src/tokenizer.rs
@@ -899,8 +899,7 @@ fn advance_symbol(code: &str) -> Advancement {
 
 #[inline]
 fn parse_whitespace(code: &str) -> IResult<&str, &str> {
-    // From regex "^[^\S\n]+\n?|\n"
-    alt((recognize(pair(opt(line_ending), space1)), line_ending))(code)
+    alt((recognize(pair(space1, opt(line_ending))), line_ending))(code)
 }
 
 // Keep finding whitespace until the line ends
@@ -1201,11 +1200,11 @@ mod tests {
     #[test]
     fn test_advance_whitespace() {
         test_advancer!(
-            advance_whitespace("\t  \n"),
+            advance_whitespace("\t  \n\t"),
             Ok(Some(TokenAdvancement {
-                advance: 3,
+                advance: 4,
                 token_type: TokenType::Whitespace {
-                    characters: Cow::from("\t  "),
+                    characters: Cow::from("\t  \n"),
                 },
             }))
         );
@@ -1223,9 +1222,19 @@ mod tests {
         test_advancer!(
             advance_whitespace("\t\t\nhello"),
             Ok(Some(TokenAdvancement {
-                advance: 2,
+                advance: 3,
                 token_type: TokenType::Whitespace {
-                    characters: Cow::from("\t\t"),
+                    characters: Cow::from("\t\t\n"),
+                },
+            }))
+        );
+
+        test_advancer!(
+            advance_whitespace("\n\thello"),
+            Ok(Some(TokenAdvancement {
+                advance: 1,
+                token_type: TokenType::Whitespace {
+                    characters: Cow::from("\n"),
                 },
             }))
         );